fix undefined behavior: lineage contains pointer to itself
The Lineage would sometimes store a pointer to a value stored inline in itself. This was intended, although I forgot to consider what would happen if the Lineage instance gets moved to a different memory address. This move cannot happen while the Lineage is borrowed, so the public api is fine (specifically Lineage::get), but the move can happen while it is not borrowed. To remedy this we will not store a pointer to a field within the Lineage itself. Instead we will store a memory offset so that we can cheaply construct the pointer whenever Lineage::get is called.

We now need two atomics. The old atomic "self.current.1" contains the memory address of a Box<T> or null. If it contains null we need to compute a pointer to somewhere within the lineage by offsetting the memory address of self by "self.current.0".